+++ app/src/main/java/com/example/disastron/FacebookShare.java
@@ -4,7 +4,7 @@ import androidx.appcompat.app.AppCompatActivity;
 
 import android.os.Bundle;
 
-public class FacebookShare extends AppCompatActivity {
+public class FacebookShare extends MainActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
