+++ app/src/main/java/com/example/disastron/FirstAid.java
@@ -4,7 +4,7 @@ import androidx.appcompat.app.AppCompatActivity;
 
 import android.os.Bundle;
 
-public class FirstAid extends AppCompatActivity {
+public class FirstAid extends MainActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
